+++ src/py/flwr/client/client_app.py
@@ -15,15 +15,16 @@
 """Flower ClientApp."""
 
 
-from typing import Callable, List, Optional
+from typing import Callable, List, Optional, Union
 
+from flwr.client.client import Client
 from flwr.client.message_handler.message_handler import (
     handle_legacy_message_from_msgtype,
 )
 from flwr.client.mod.utils import make_ffn
-from flwr.client.typing import ClientFnExt, Mod
+from flwr.client.typing import ClientFn, ClientFnExt, Mod
 from flwr.common import Context, Message, MessageType
-from flwr.common.logger import warn_preview_feature
+from flwr.common.logger import warn_deprecated_feature, warn_preview_feature
 
 from .typing import ClientAppCallable
 
@@ -35,6 +36,31 @@ class ClientAppException(Exception):
         ex_name = self.__class__.__name__
         self.message = f"\nException {ex_name} occurred. Message: " + message
         super().__init__(self.message)
+
+
+def _inspect_maybe_adapt_client_fn_signature(
+    client_fn: Union[ClientFn, ClientFnExt]
+) -> ClientFnExt:
+
+    if "cid" in client_fn.__annotations__:
+        warn_deprecated_feature(
+            "Passing a `client_fn` with signature `def client_fn(cid: str)` "
+            "is deprecated. Use instead signature `def client_fn(node_id: int, "
+            "partition_id: Optional[int])`.",
+        )
+
+        # Wrap depcreated client_fn inside a function with the expected signature
+        def adaptor_fn(
+            node_id: int, partition_id: Optional[int]  # pylint: disable=unused-argument
+        ) -> Client:
+            return client_fn(str(partition_id))  # type: ignore
+
+    else:
+
+        def adaptor_fn(node_id: int, partition_id: Optional[int]) -> Client:
+            return client_fn(node_id, partition_id)  # type: ignore
+
+    return adaptor_fn
 
 
 class ClientApp:
@@ -72,7 +98,9 @@ class ClientApp:
 
         # Create wrapper function for `handle`
         self._call: Optional[ClientAppCallable] = None
-        self.client_fn = client_fn
+        self.client_fn = (
+            _inspect_maybe_adapt_client_fn_signature(client_fn) if client_fn else None
+        )
 
         # Step functions
         self._train: Optional[ClientAppCallable] = None
